requirements.txt
autopep8==1.5.7
cffi==1.14.6
cryptography==35.0.0
distro==1.6.0
httplib2==0.20.1
importlib-metadata==4.8.1
jeepney==0.7.1
keyring==23.2.1
launchpadlib==1.10.14
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
mysql-connector-python==8.0.26
mysqlclient==2.0.3
oauthlib==3.1.1
pbr==5.6.0
pkg_resources==0.0.0
protobuf==3.18.1
pycodestyle==2.7.0
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
SecretStorage==3.3.1
six==1.16.0
testresources==2.0.1
toml==0.10.2
wadllib==1.3.6
zipp==3.6.0